fix: Template variables are not applied on client
